Module serde_hex::config

SerHex configuration variants.

This module is a collection of marker types which implement the possible combinations of config values described by the HexConf trait. All config values are supplied with associated functions that are marked with #[inline]. This ensures that the compiler will (usually) optimize away all configuration checks.

Structs

Compact

Config indicating compact representation with no capitalization and no prefixing.

CompactCap

Config indicating compact representation with capitalization but no prefixing.

CompactCapPfx

Config indicating compact representation with capitalization and prefixing.

CompactPfx

Config indicating compact representation with prefixing but no capitalization.

Strict

Config indicating a strict representation with no capiltaization and no prefixing.

StrictCap

Config indicating a strict representation with capitalization but no prefixing.

StrictCapPfx

Config indicating a strict representation with capitalization and prefixing.

StrictPfx

Config indicating a strict representation with prefixing but no capitalization.

Traits

HexConf

Trait for supplying configuration to SerHex. This trait takes no self parameters, as it is intended to be applied unit structs. All default implementation are set to false.